🐛 fix     :  통지 연결 요청 버그
 통지 연결끊어질시 다시 자동으로 연결요청

+++ src/main/java/lcw/lcw2_back/service/notification/NotificationServiceImpl.java
@@ -12,9 +12,6 @@ import org.springframework.web.servlet.mvc.method.annotation.SseEmitter;
 import java.io.IOException;
 import java.util.List;
 import java.util.Map;
-import java.util.concurrent.Executors;
-import java.util.concurrent.ScheduledExecutorService;
-import java.util.concurrent.TimeUnit;
 
 @Service
 @AllArgsConstructor
@@ -25,31 +22,21 @@ public class NotificationServiceImpl implements NotificationService{
 
     @Override
     public SseEmitter connectSSE(String userId) {
+        System.out.println("연결시도중....");
         String emitterId = userId;
-        SseEmitter emitter = emitterRepository.save(emitterId,new SseEmitter(timeout));
-
-        ScheduledExecutorService executor = Executors.newScheduledThreadPool(1);
-
-        // 주기적(30초)으로 더미 이벤트 전송
-        executor.scheduleAtFixedRate(() -> {
-            try {
-                emitter.send(SseEmitter.event().data("dummy")); // 더미 데이터 전송
-            } catch (Exception e) {
-                emitter.completeWithError(e);
-            }
-        }, 0, 30, TimeUnit.SECONDS); // 처음 0초 후 시작하고, 30초마다 실행
-
-        // SSE 연결이 끝나면 스케줄러 종료
-        emitter.onCompletion(()->{
-            executor.shutdown();
-            updateNotificationTableByUserId(userId);
-            emitterRepository.deleteById(emitterId);
-            //로그아웃이나 타임아웃시에 디비접근해서 한꺼번에 읽음 처리
-        });
-        emitter.onTimeout(()->{
-            updateNotificationTableByUserId(userId);
-            emitterRepository.deleteById(emitterId);
-        });
+        SseEmitter emitter = emitterRepository.findEmitterByUserId(userId);
+        if(emitter == null) {
+            emitter = emitterRepository.save(emitterId, new SseEmitter(-1L));
+            emitter.onCompletion(()->{
+                updateNotificationTableByUserId(userId);
+                emitterRepository.deleteById(emitterId);
+                //로그아웃이나 타임아웃시에 디비접근해서 한꺼번에 읽음 처리
+            });
+            emitter.onTimeout(()->{
+                updateNotificationTableByUserId(userId);
+                emitterRepository.deleteById(emitterId);
+            });
+        }
         updateEventCacheFromNotificationDB(userId);
         //연결시 통지들 cache에 들고온다.
 
@@ -95,7 +82,7 @@ public class NotificationServiceImpl implements NotificationService{
         Notification notification= notificationRepository.save(new Notification(eventId,receivedUserId, notificationType.name(), content,notificationType));
 
 
-        SseEmitter emitter = emitterRepository.findEmitterByUserId(String.valueOf(receivedUserId));
+        SseEmitter emitter = emitterRepository.findEmitterByUserId(receivedUserId);
         emitterRepository.saveEventCache(eventId,notification);
         if(emitter == null){
             System.out.println("상대방이 로그아웃인 상태이거나 받을 수 없습니다.");
